# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: 0.0.1
  title: Trivia OFB
# during dev, should point to your local machine
host: localhost:10011
schemes:
  - http
consumes:
  - application/json
  - multipart/form-data
produces:
  - application/json
basePath: /trivia


definitions:
  
  Credentials:
    type: object
    properties:
      userName:
        type: string
      password:
        type: string
  
  ResponseMessage:
    type: object
    properties:
      messeage:
        type: string
  
  Admin:
    type: object
    properties:
      email:
        type: string
      name:
        type: string
      lastName:
        type: string
      userName:
        type: string
      password:
        type: string
      active:
        type: string
  
  AnswerInfo:
    type: object
    properties:
      content:
        type: string
      date:
        type: string
        format: date-time
      userId: 
        type: integer
      score: 
        type: number
        format: float
      winner:
        type: integer
  
  User:
    type: object
    properties:
      id: 
        type: integer
      name:
        type: string
      lastName:
        type: string
      email:
        type: string
      cellphone:
        type: string
      contactPreference:
        type: string
  
  WinnerInfo:
    type: object
    properties:
      questionId:
        type: integer
      userId:
        type: integer
  
  Video:
    type: object
    properties:
      id:
        type: integer
      title:
        type: string
      channel:
        type: string
      thumbnail:
        type: string
      url:
        type: string
      time:
        type: string
      selected:
        type: boolean
      active:
        type: boolean

  VideoData:
    type: object
  
  Audio:
    type: object
    properties:
      id:
        type: integer
      title:
        type: string
      artist:
        type: string
      url:
        type: string
      selected:
        type: boolean
      active:
        type: boolean
  
  Question:
    type: object
    properties:
      id:
        type: integer
      content:
        type: string
      createdDate:
        type: string
        format: date
      startDate:
        type: string
        format: date
      endDate:
        type: string
        format: date
      answer:
        type: string
      status:
        type: string
      active:
        type: boolean
      
  
paths:  

  /get-question:
    x-swagger-router-controller: questionController
    
    get:
      operationId: getQuestion
      description: Retrieves the question of the trivia and the date that it ends.
      tags:
        - questions
      
      parameters:
        - in: query
          name: id
          type: integer
      
      responses:
        
        200:
          description: A JSON object with question and timeToEnd
          schema:
            $ref: "#/definitions/Question"
        
        default:
          description: Something went wrong.
          schema:
            $ref: "#/definitions/User"

  
  /get-user:
    x-swagger-router-controller: citizenController
    
    get:
      operationId: getUser
      description: Retrieves the especifed user
      tags:
        - admins
      
      parameters:
        - in: query
          name: id
          type: integer
      
      responses:
        
        200:
          description: ok
          schema:
            $ref: "#/definitions/User"
        
        default:
          description: Something went wrong.
          schema:
            $ref: "#/definitions/User"
  
  /login:
    x-swagger-router-controller: adminController
    
    post:
      operationId: login
      description: login for admin
      tags:
        - admins
    
      parameters:
        - in: body
          name: credentials
          required: true
          schema:
            $ref: "#/definitions/Credentials"
      
      responses:
        
        200:
          description: logued in
          schema:
            $ref: "#/definitions/ResponseMessage"
        
        401:
          description: wrong credentials
          schema:
            $ref: "#/definitions/ResponseMessage"
        
        default:
          description: wrong credentials
          schema:
            $ref: "#/definitions/ResponseMessage"
  
  /register:
    x-swagger-router-controller: adminController
    
    post:
      operationId: registerAdmin
      description: registers a new admin
      tags:
        - admins
      
      parameters:
        - in: body
          name: admin
          schema:
            $ref: "#/definitions/Admin"
      
      responses:
        
        201:
          description: ok
          schema:
            $ref: "#/definitions/ResponseMessage"
        
        default:
          description: An error ocurred
          schema:
            $ref: "#/definitions/ResponseMessage"
  
  /users-answers:
    x-swagger-router-controller: answerController
  
    get:
      operationId: getAnswersList
      description: Retrieves the answers from the users for the given question
      tags:
        - answers
      
      parameters:
        - in: query
          name: questionId
          type: integer
          required: true
      
      responses:
        
        200:
          description: ok
          schema:
            type: array
            items:
              $ref: "#/definitions/AnswerInfo"
        
        default:
          description: Something went wrong.
          schema:
            $ref: "#/definitions/User"
  
  /questions-list:
    x-swagger-router-controller: questionController
  
    get:
      operationId: getQuestionsList
      description: Retrieves the answers from the users for the given question
      tags:
        - questions
      
      parameters:
        - in: query
          name: lastId
          type: integer
        - in: query
          name: page
          type: string
      
      responses:
        
        200:
          description: ok
          schema:
            type: array
            items:
              $ref: "#/definitions/Question"
        
        default:
          description: error
          schema:
            $ref: "#/definitions/ResponseMessage"
  
  /select-winners:
    x-swagger-router-controller: answerController
    
    post:
      operationId: selectWinners
      summary: Sets useras answers as winners.
      tags:
        - answers
      
      parameters:
        - in: body
          name: winners
          required: true
          schema:
            type: array
            items:
              $ref: "#/definitions/WinnerInfo"
      
      responses:
      
        200:
          description: ok
          schema:
            $ref: "#/definitions/ResponseMessage"
        
        default:
          description: Something went wrong
          schema:
            $ref: "#/definitions/ResponseMessage"
  
  /get-video-data:
    x-swagger-router-controller: videoController
    
    get:
      operationId: getVideoData
      summary: An object with content and time to end.
      description: Retrieves the question of the trivia and the date that it ends.
      tags:
        - videos
      
      parameters:
        - in: query
          name: url
          type: string

      responses:
        
        200:
          description: A JSON object with question and timeToEnd
          schema:
            $ref: "#/definitions/VideoData"
        
        default:
          description: Something went wrong.
          schema:
            $ref: "#/definitions/User"

  /upload-video:
    x-swagger-router-controller: videoController
    
    post:
      operationId: manageVideoData
      description: Creates or updates a video
      tags:
        - videos
      
      parameters:
        - in: body
          name: video
          schema:
            $ref: "#/definitions/Video"
      
      responses:
        
        200:
          description: video updated
          schema:
            $ref: "#/definitions/Video"
        
        201:
          description: Video created
          schema:
            $ref: "#/definitions/Video"
        
        default:
          description: Something went wrong.
          schema:
            $ref: "#/definitions/ResponseMessage"
  
  /get-videos:
    x-swagger-router-controller: videoController
    
    get:
      operationId: getVideos
      description: Retrieves the five videos
      tags:
        - videos
      
      responses:
        
        200:
          description: video updated
          schema:
            type: array
            items:
              $ref: "#/definitions/Video"
        
        default:
          description: Something went wrong.
          schema:
            $ref: "#/definitions/ResponseMessage"
  
  /get-video:
    x-swagger-router-controller: videoController
    
    get:
      operationId: getVideo
      description: Retrieves the specified video
      tags:
        - videos
      
      parameters:
        - in: query
          name: id
          type: integer
      
      responses:
        
        200:
          description: ok
          schema:
            $ref: "#/definitions/Video"
        
        default:
          description: Something went wrong.
          schema:
            $ref: "#/definitions/ResponseMessage"
  
  /get-audios:
    x-swagger-router-controller: audioController
    
    get:
      operationId: getAudios
      description: Retrieves the five audios
      tags:
        - audios
      
      responses:
        
        200:
          description: ok
          schema:
            type: array
            items:
              $ref: "#/definitions/Audio"
        
        default:
          description: Something went wrong.
          schema:
            $ref: "#/definitions/ResponseMessage"
  
  /get-audio:
    x-swagger-router-controller: audioController
    
    get:
      operationId: getAudio
      description: Retrieves the especified audio
      tags:
        - audios
      
      parameters:
        - in: query
          name: id
          type: integer
      
      responses:
        
        200:
          description: ok
          schema:
            $ref: "#/definitions/Audio"
        
        default:
          description: Something went wrong.
          schema:
            $ref: "#/definitions/ResponseMessage"
  
  /upload-audio:
    x-swagger-router-controller: audioController
    
    post:
      operationId: manageAudio
      description: Creates or updates an audio link
      tags:
        - audios
      
      parameters:
        - in: body
          name: audio
          schema:
            $ref: "#/definitions/Audio"
      
      responses:
        
        200:
          description: Audio updated
          schema:
            $ref: "#/definitions/Audio"
        
        201:
          description: Audio created
          schema:
            $ref: "#/definitions/Audio"
        
        default:
          description: Something went wrong.
          schema:
            $ref: "#/definitions/ResponseMessage"
  
  /upload-question:
    x-swagger-router-controller: questionController
    
    post:
      operationId: manageQuestion
      description: creates a new question or updates a new one
      tags:
        - questions
      
      parameters:
        - in: body
          name: question
          schema:
            $ref: "#/definitions/Question"
      
      responses:
        
        200:
          description: ok
          schema:
            $ref: "#/definitions/Question"
        
        201:
          description: ok
          schema:
            $ref: "#/definitions/Question"
        
        default:
          description: Something went wrong.
          schema:
            $ref: "#/definitions/ResponseMessage"